This research project is aimed to propose an action-elimination algorithm in an attempt to learn the Implicit time windows in a Travelling Salesman Problem with a time-window setting.
In a parcel-delivery context, implicit time windows can be visualized as the local knowledge possessed by a delivery driver regarding various factors influencing the total travelling time of a trip. Learning the implicit time windows allows dispatchers to adaptively route customers in a sequence that closely resembles a sequence that drivers may adopt. To achieve this objective, we developed ‘Jarvis the Eliminator’, a constraint programming powered solver that is able to suggest time-bucket solutions based on a priori knowledge acquired for every customer in a particular route. By formulating a driver’s route as a TSP-TW instance and a naïve route as a simple TSP, we employ Jarvis the Eliminator to iteratively reduce the solution space and suggest time-buckets for the naïve route solution.  Dataset Used: Solomon's RC2 dataset with a single vehicle, originally created by Potvin & Bengio (1996)